docs: add 3gpp compliance page

+++ internal/upf/engine/pdr.go
@@ -30,5 +30,9 @@ func unapplyPDR(spdrInfo SPDRInfo, bpfObjects *ebpf.BpfObjects) error {
 		return bpfObjects.DeletePdrDownlink(spdrInfo.UEIP)
 	}
 
-	return bpfObjects.DeletePdrUplink(spdrInfo.TeID)
+	if spdrInfo.TeID != 0 {
+		return bpfObjects.DeletePdrUplink(spdrInfo.TeID)
+	}
+
+	return nil
 }
